divide implementation to smaller pieces
